Proposed method to force inclusion of appliction.js.
I'm hoping I just don't understand some of the loading and configuration. I've uncommented

	require application

in the spec_helper.coffee, and the require appears in the browser inspector, but does
not seem to work. This little patch to the runner view fixes both the web page and
headless run for me.