Removes thread check as the thread may be dead
Depending on the rack_options passed into Rack::Server the thread may
not stay running and can simply exit immediately. If that happens then
this will always return false. I doesn't seem to be necessary to check
the thread status before checking if the server is running via TCP so
I've removed this check.