Do not double load teaspoon_env
When loading from the console, we need to load the Rails env, so we
load the app's teaspoon_env. When we load teaspoon_env, the app then
loads all engines, including teaspoon. Previously, this double loaded
teaspoon_env and caused additive effects on configurations that
accumulate values, like #hooks. (Fixes #332)