Rework Sprockets 4 support to work with a CoffeeScript spec helper
With Sprockets 4 every asset is associated with a MIME type and if none
is provided then it is inferred from the file name. Having a
CoffeeScript spec helper will implicitly set the MIME type to
'text/coffeescript' which will be passed on to any dependencies,
resulting in missing JavaScript assets. However by explicity setting
the acceptable MIME type to 'application/javascript' we can use a
CoffeeScript spec helper while also having JavaScript dependencies.

ALso, when we are not running coverage tests we want the debug assets as
is the default with Sprockets 4 in a non-production environment, giving
us the advantage of source maps. When running coverage tests we do not
serve debug assets but rather 'self' assets since debug assets in
Sprockets 4 concatenate assets included by Sprockets directives.