passing --filter="...." correctly escapes the spaces into "%20" and not "+"
Running teaspoon as

bundle exec teaspoon -s default --filter="ScenePointers are attached"

When using CGI.escape the string "ScenePointers are attached" becomes "ScenePointesr+are+attached"
This is then passed as a "grep=ScenePointers+are+attached" into the url

In the JS that is running the specs in the browser the value of the grep is used to filter.
We move through all the specs and encode their full name with encodeURIComponent

The result is

encodeURIComponent("ScenePointers are added")
'ScenePointers%20are%20added'

This means that on the ruby side we were escaping the space " " in the urls as '+' and on the
JS side we were escaping them as "%20". One of these had to change and it seems changing it on
the ruby side is the right one.

There is no right way to escape. We can use 'addressable' and URI and CGI and ERB.
ERB seems to be the most straighforward for this case, it also does not bring a lot of dependencies.
Just one - 'cgi'.

In the same time there seems to be no method in URI or CGI or 'addressable' gem that would address this